+++ tiledb/sm/serialization/query.cc
@@ -45,6 +45,11 @@
 #include "tiledb/sm/query/query.h"
 #include "tiledb/sm/query/reader.h"
 #include "tiledb/sm/query/writer.h"
+
+#ifdef _WIN32
+#include "tiledb/sm/serialization/meet-capnproto-win32-include-expectations.h"
+#endif
+
 #include "tiledb/sm/serialization/capnp_utils.h"
 #include "tiledb/sm/subarray/subarray.h"
 #include "tiledb/sm/subarray/subarray_partitioner.h"